Post-Merge fix, refactoring split inlineDrawer
