+++ src/main/java/org/overengineer/inlineproblems/InlineDrawer.java
@@ -2,137 +2,26 @@ package org.overengineer.inlineproblems;
 
 import com.intellij.lang.annotation.HighlightSeverity;
 import com.intellij.openapi.Disposable;
-import com.intellij.openapi.application.ApplicationInfo;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.*;
 import com.intellij.openapi.editor.colors.EditorFontType;
 import com.intellij.openapi.editor.markup.HighlighterTargetArea;
 import com.intellij.openapi.editor.markup.MarkupModel;
 import com.intellij.openapi.editor.markup.TextAttributes;
-import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.TextRange;
 import org.overengineer.inlineproblems.entities.InlineProblem;
-import org.overengineer.inlineproblems.entities.enums.Listeners;
 import org.overengineer.inlineproblems.settings.SettingsState;
 
 import java.awt.*;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.stream.Collectors;
 
 
-public class InlineDrawer implements Disposable {
-    private List<InlineProblem> activeProblems = new ArrayList<>();
+public class InlineDrawer {
 
-    private Logger logger = Logger.getInstance(InlineDrawer.class);
-    private ListenerManager listenerManager = new ListenerManager();
-
-    private static final String RIDER_NAME = "JetBrains Rider";
-
-    public InlineDrawer() {
-        if (ApplicationInfo.getInstance().getFullApplicationName().startsWith(RIDER_NAME)) {
-            // todo unity detection
-        }
-    }
-
-    public void dispose() {
-        reset();
-    }
-
-    public void removeProblem(InlineProblem problem) {
-        if (problem.getProblemInLineCount() > 0) {
-            problem.setProblemInLineCount(problem.getProblemInLineCount() - 1);
-            return;
-        }
-
-        undrawErrorLineHighlight(problem);
-        undrawInlineProblemLabel(problem);
-        activeProblems.remove(problem);
-    }
-
-    public boolean removeProblemWithRefreshFromActiveProblems(InlineProblem problem) {
-        if (problem.getProblemInLineCount() > 0) {
-            problem.setProblemInLineCount(problem.getProblemInLineCount() - 1);
-            return true;
-        }
-
-        final List<InlineProblem> activeProblemSnapShot = new ArrayList<>(activeProblems);
-
-        List<InlineProblem> problemsToRemove = activeProblemSnapShot.stream()
-                        .filter(p -> p.equals(problem))
-                                .collect(Collectors.toList());
-
-        if (problemsToRemove.size() != 1) {
-            logger.info("Problem to remove not found, resetting");
-            reset();
-            return false;
-        }
-
-        for (var problemToRemove : problemsToRemove) {
-            removeProblem(problemToRemove);
-        }
-
-        return true;
-    }
-
-    public void addProblem(InlineProblem problem) {
-        AtomicBoolean problemExistsInLine = new AtomicBoolean(false);
-
-        activeProblems.stream()
-                .filter(p -> p.getText().equals(problem.getText()) && p.getLine() == problem.getLine())
-                .findFirst()
-                .ifPresent(p -> {
-                    p.setProblemInLineCount(p.getProblemInLineCount() + 1);
-                    problemExistsInLine.set(true);
-                });
-
-        if (problemExistsInLine.get())
+    public void drawProblemLabel(InlineProblem problem) {
+        if (!shouldDrawProblemLabel(problem))
             return;
 
-        if (shouldDrawProblemLabel(problem))
-            drawProblemLabel(problem);
-
-        if (shouldDrawProblemHighlighter(problem))
-            drawProblemLineHighlight(problem);
-
-        activeProblems.add(problem);
-    }
-
-    public void reset(int enabledListener, boolean listenerChanged) {
-        reset();
-
-        if (listenerChanged && enabledListener == Listeners.MARKUP_MODEL_LISTENER) {
-            listenerManager.installMarkupModelListenerOnAllProjects();
-        }
-    }
-
-    public void reset() {
-        final List<InlineProblem> activeProblemSnapShot = new ArrayList<>(activeProblems);
-        activeProblemSnapShot.forEach(this::removeProblem);
-    }
-
-    public void updateFromListOfNewActiveProblems(List<InlineProblem> problems, Project project, String filePath) {
-
-        if (problems.size() == 0 && problems == activeProblems)
-            return;
-
-        final List<InlineProblem> processedProblems = new ArrayList<>();
-        final List<InlineProblem> activeProblemsSnapShot = activeProblems.stream()
-                .filter(p -> p.getProject().equals(project) && p.getFile().equals(filePath))
-                .collect(Collectors.toList());
-
-        activeProblemsSnapShot.stream()
-                .filter(p -> !problems.contains(p))
-                .forEach(p -> {processedProblems.add(p); removeProblem(p);});
-
-        problems.stream()
-                .filter(p -> !activeProblemsSnapShot.contains(p) && !processedProblems.contains(p))
-                .forEach(this::addProblem);
-    }
-
-    private void drawProblemLabel(InlineProblem problem) {
         Editor editor = problem.getEditor();
         SettingsState settings = SettingsState.getInstance();
         var inlayModel = editor.getInlayModel();
@@ -192,7 +81,10 @@ public class InlineDrawer implements Disposable {
         problem.setInlineProblemLabelHashCode(inlineProblemLabel.hashCode());
     }
 
-    private void drawProblemLineHighlight(InlineProblem problem) {
+    public void drawProblemLineHighlight(InlineProblem problem) {
+        if (!shouldDrawProblemHighlighter(problem))
+            return;
+
         Editor editor = problem.getEditor();
         TextAttributes textAttributes = new TextAttributes(
                 editor.getColorsScheme().getDefaultForeground(),
@@ -213,7 +105,7 @@ public class InlineDrawer implements Disposable {
         ).hashCode());
     }
 
-    private void undrawErrorLineHighlight(InlineProblem problem) {
+    public void undrawErrorLineHighlight(InlineProblem problem) {
         MarkupModel markupModel = problem.getEditor().getMarkupModel();
 
         Arrays.stream(markupModel.getAllHighlighters())
@@ -221,7 +113,7 @@ public class InlineDrawer implements Disposable {
                 .forEach(markupModel::removeHighlighter);
     }
 
-    private void undrawInlineProblemLabel(InlineProblem problem) {
+    public void undrawInlineProblemLabel(InlineProblem problem) {
         Editor editor = problem.getEditor();
         Document document = editor.getDocument();
